feat: web-wagmi restructure
