+++ examples/web-wagmi/src/profiles/components/ProfileFollowers.tsx
@@ -1,6 +1,6 @@
 import { useProfileFollowers } from '@lens-protocol/react';
 
-import { Loading } from '../loading/Loading';
+import { Loading } from '../../components/loading/Loading';
 import { ProfileCard } from './ProfileCard';
 
 type ProfileFollowersProps = {
