+++ examples/web-wagmi/src/profiles/components/ProfileFollowing.tsx
@@ -1,6 +1,6 @@
 import { useProfileFollowing } from '@lens-protocol/react';
 
-import { Loading } from '../loading/Loading';
+import { Loading } from '../../components/loading/Loading';
 import { ProfileCard } from './ProfileCard';
 
 type ProfileFollowingProps = {
