+++ examples/web-wagmi/src/profiles/components/ProfilePicture.tsx
