+++ examples/web-wagmi/src/publications/UseCreatePost.tsx
@@ -1,7 +1,7 @@
 import { useActiveProfile } from '@lens-protocol/react';
 
 import { LoginButton } from '../components/auth/LoginButton';
-import { CreatePost } from '../components/feed/CreatePost';
+import { CreatePost } from './components/CreatePost';
 
 export function UseCreatePost() {
   const { data: profile } = useActiveProfile();
